+++ src/main/java/com/fairydeck/task/model/request/AddJobRequest.java
@@ -2,10 +2,12 @@ package com.fairydeck.task.model.request;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class AddJobRequest {
     private int min = 1;
     private int max = 8;
-    private int limit = 1;
+    private BigInteger limit = BigInteger.valueOf(1);
     private String charMap = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
 }
